requirements.txt
Flask
torchvision
requests
Pillow
torch
Pillow